Added support for models to have an initial transformation, and they have an initial and current transformation, invoking reset sets current TraFo intial TraFo.
⭐⭐⭐ With that the model refactor is complete as we have functional UBO ⭐⭐⭐

Optimization would be to have shared memory for the camera position and delta time

the next step is to modularize textures so that we can different models and different textures. The model class should not only have a mesh, initial and current transformation but also material information. Maybe script / function callbacks, but that's for later. Once materials are in we can create models however we want. To make it simple for now a material is simply a texture. Maybe I'll remove teh seperate between diffrent types of models so every material would also have color entries, we'll see.